Infra/Pr02/ejercicio 8/backend/requirements.txt
flask==2.3.0
psycopg2-binary==2.9.6
redis==4.5.0
